final determination on Require statements in config
This declaration is coming from a fresh install.
`<FilesMatch>` did not work and would prefer not `<Location>`
so that it can be overridden easier. That being said,
Location may need to be granted if other config denies
the webspace path.

The removal for `<RequireAll>` is due to Apache joining
each file's overridding config in a `<RequireAny>`,
effectively making this redundant.