Token Endpoint WITHOUT Scopes
MinToken requires *some* scope when it passes off auth to SelfAuth
and processes the result (adding token). IndieAuth-Client-php
does not, by default, use the token endpoint when there are no
scopes; however, this may be outdated and other tools (such as
IndieLogin.com) do. This change makes it compatible with these tools.